requirements.txt
Flask
edge_tts
playsound
google-generativeai
python-dotenv
asyncio
baymaxtts
requests
```